+++ test_renderer.py
@@ -371,6 +371,51 @@ def test_renderer_same_str():
 
     assert str(h1) == str(h2)
 
+
+def test_discovering_js():
+    class O(Tag):
+        js="/*JS1*/"
+
+    # dynamic js init not present (normal!)
+    assert "/*JS1*/" not in str(HRenderer( O(), "//"))
+
+    class OOI(Tag): # immediate rendering
+        def __init__(self):
+            Tag.__init__(self)
+            self.set( O() )             # Tag directly in Tag
+
+    class OOOI(Tag):  # immediate rendering
+        def __init__(self):
+            Tag.__init__(self)
+            self.set( Tag.div( O() ) )  # Tag in a TagBase
+
+    class OOL(Tag):   # lately rendering
+        def __str__(self):
+            self.set( O() )             # Tag directly in Tag
+            return Tag.__str__(self)
+
+    class OOOL(Tag):  # lately rendering
+        def __str__(self):
+            self.set( Tag.div(O()) )    # Tag in a TagBase
+            return Tag.__str__(self)
+
+    async def test(r): # first call (init obj)
+        resp = await r.interact( 0, None, None, None)
+        assert resp["update"][0]["id"] == 0
+        assert "/*JS1*/" in resp["post"]
+
+    r=HRenderer(OOI(),"//js interact")
+    asyncio.run(test(r))
+
+    r=HRenderer(OOOI(),"//js interact")
+    asyncio.run(test(r))
+
+    r=HRenderer(OOL(),"//js interact")
+    asyncio.run(test(r))
+
+    r=HRenderer(OOOL(),"//js interact")
+    asyncio.run(test(r))
+
 # test_ko_try_render_a_tagbase()
 # test_render_a_tag_with_interaction()
 # test_render_a_tag_with_child_interactions()